test(US-005): split debate test file and add precise cost tracking tests
- Extract shared helpers to _rectification-debate-helpers.ts
- Split rectification-loop-debate.test.ts (disabled/enabled/fallback) to stay under 400 lines
- Add rectification-loop-debate-cost.test.ts with injectable runDebate mock
- New tests verify story.routing.estimatedCost mutation when totalCostUsd > 0
- New tests verify cost is unchanged when totalCostUsd === 0
